@@ -1,6 +1,6 @@
 {
   "name": "react-native-pspdfkit",
-  "version": "1.24.9",
+  "version": "1.25.0",
   "description": "A React Native module for the PSPDFKit library.",
   "keywords": [
     "react native",
@@ -23,8 +23,8 @@
   },
   "peerDependencies": {
     "prop-types": ">= 15.6.2 < 16",
-    "react": ">= 16.5.0 < 17",
-    "react-native": ">= 0.57.2 < 1"
+    "react": ">= 16.8.6 < 17",
+    "react-native": ">= 0.60.4 < 1"
   },
   "dependencies": {},
   "devDependencies": {
