@@ -1,6 +1,6 @@
 package com.pspdfkit.react.events;
 
-import android.support.annotation.IdRes;
+import androidx.annotation.IdRes;
 
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.WritableMap;
