@@ -1,7 +1,7 @@
 package com.pspdfkit.react.helper;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.pspdfkit.document.providers.DataProvider;
 
