@@ -1,8 +1,8 @@
 package com.pspdfkit.views;
 
 import android.graphics.PointF;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.MotionEvent;
 
 import com.facebook.react.uimanager.events.EventDispatcher;
